refactor: refine code
- remove unused file
- upgrade to 0.1.2
