+++ solagent-adapters/rig/pumpfun/.gitignore
@@ -1,0 +1,1 @@
+target/
